Save changes

+++ datasetGenerator_v2.h
@@ -7,10 +7,10 @@ class DatasetGenerator {
 public:
     DatasetGenerator();
 
-    std::vector<int> generateOrdered(int length);
-    std::vector<int> generateOrderedInverse(int length);
-    std::vector<int> generateAlmostOrdered(int length);
-    std::vector<int> generateRandom(int length);
+    std::vector<long long> generateOrdered(long long length);
+    std::vector<long long> generateOrderedInverse(long long length);
+    std::vector<long long> generateAlmostOrdered(long long length);
+    std::vector<long long> generateRandom(long long length);
 
 private:
 };
